Support async code + add Armada.readFile API
To support async I/O, we need to have an event loop that can drive
Futures in Rust, and the JS side event loop in V8. For this we use
Tokio.

Firstly, we change the main function in Rust to an async function and
specify Tokio as the async runtime to use. We then configure deno_core
to run the JS event loop with the Tokio async runtime.

Next we add a new async op that uses `tokio::fs` to read a file from
disk. We then add a nice JS API that wraps this op.

We test these changes by asynchronously reading a file in example.js.