Load the deno_console extension
This adds a `globalThis.console` implementation to the runtime. The code
backing this implementation is contained within the `deno_console`
extension. We load this extension into the runtime, and then initialize
the `globalThis.console` global in our bootstrap code. The console
prints to stdout/stderr using the `print` function from deno_core.